+++ src/Plugin/Wechat/V3/Marketing/Transfer/InvokeIosPlugin.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\MchTransfer;
+namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer;
 
 use Closure;
 use Yansongda\Artful\Contract\PluginInterface;
@@ -21,9 +21,9 @@ use function Yansongda\Pay\get_provider_config;
 use function Yansongda\Pay\get_wechat_type_key;
 
 /**
- * @see https://pay.weixin.qq.com/doc/v3/merchant/4012719576
+ * @see https://pay.weixin.qq.com/doc/v3/merchant/4012719578
  */
-class InvokeAndroidPlugin implements PluginInterface
+class InvokeIosPlugin implements PluginInterface
 {
     /**
      * @throws ContainerException
@@ -36,20 +36,20 @@ class InvokeAndroidPlugin implements PluginInterface
         /* @var Rocket $rocket */
         $rocket = $next($rocket);
 
-        Logger::debug('[Wechat][V3][Marketing][MchTransfer][InvokeAndroidPlugin] 插件开始装载', ['rocket' => $rocket]);
+        Logger::debug('[Wechat][V3][Marketing][Transfer][InvokeIosPlugin] 插件开始装载', ['rocket' => $rocket]);
 
         $config = get_provider_config('wechat', $rocket->getParams());
         $destination = $rocket->getDestination();
         $packageInfo = $destination?->get('package_info');
 
         if (Pay::MODE_SERVICE === ($config['mode'] ?? Pay::MODE_NORMAL)) {
-            throw new InvalidParamsException(Exception::PARAMS_PLUGIN_ONLY_SUPPORT_NORMAL_MODE, '参数异常: Android调起用户确认收款，只支持普通商户模式，当前配置为服务商模式');
+            throw new InvalidParamsException(Exception::PARAMS_PLUGIN_ONLY_SUPPORT_NORMAL_MODE, '参数异常: iOS调起用户确认收款，只支持普通商户模式，当前配置为服务商模式');
         }
 
         if (is_null($packageInfo)) {
-            Logger::error('[Wechat][V3][Marketing][MchTransfer][InvokeAndroidPlugin] Android调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求', $destination?->all() ?? null);
+            Logger::error('[Wechat][V3][Marketing][Transfer][InvokeIosPlugin] iOS调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求', $destination?->all() ?? null);
 
-            throw new InvalidResponseException(Exception::RESPONSE_MISSING_NECESSARY_PARAMS, $destination?->get('fail_reason') ?? 'Android调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求', $destination?->all() ?? null);
+            throw new InvalidResponseException(Exception::RESPONSE_MISSING_NECESSARY_PARAMS, $destination?->get('fail_reason') ?? 'iOS调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求', $destination?->all() ?? null);
         }
 
         $params = $rocket->getParams();
@@ -58,7 +58,7 @@ class InvokeAndroidPlugin implements PluginInterface
 
         $rocket->setDestination($this->getInvokeConfig($payload, $params, $config, $packageInfo));
 
-        Logger::info('[Wechat][V3][Marketing][MchTransfer][InvokeAndroidPlugin] 插件装载完毕', ['rocket' => $rocket]);
+        Logger::info('[Wechat][V3][Marketing][Transfer][InvokeIosPlugin] 插件装载完毕', ['rocket' => $rocket]);
 
         return $rocket;
     }