+++ tests/Plugin/Wechat/V3/Marketing/Transfer/InvokeIosPluginTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\MchTransfer;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Transfer;
 
 use Yansongda\Artful\Exception\InvalidParamsException;
 use Yansongda\Artful\Exception\InvalidResponseException;
 use Yansongda\Artful\Rocket;
 use Yansongda\Pay\Exception\Exception;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\MchTransfer\InvokeIosPlugin;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer\InvokeIosPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;