+++ tests/Plugin/Wechat/V3/Marketing/Transfer/InvokeJsapiPluginTest.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\MchTransfer;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Transfer;
 
 use Yansongda\Artful\Exception\InvalidParamsException;
+use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidResponseException;
 use Yansongda\Artful\Rocket;
-use Yansongda\Pay\Exception\Exception;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\MchTransfer\InvokeAndroidPlugin;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer\InvokeJsapiPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class InvokeAndroidPluginTest extends TestCase
+class InvokeJsapiPluginTest extends TestCase
 {
-    protected InvokeAndroidPlugin $plugin;
+    protected InvokeJsapiPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new InvokeAndroidPlugin();
+        $this->plugin = new InvokeJsapiPlugin();
     }
 
     public function testModeWrong()
@@ -28,7 +28,7 @@ class InvokeAndroidPluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_PLUGIN_ONLY_SUPPORT_NORMAL_MODE);
-        self::expectExceptionMessage('参数异常: Android调起用户确认收款，只支持普通商户模式，当前配置为服务商模式');
+        self::expectExceptionMessage('参数异常: JSAPI调起用户确认收款，只支持普通商户模式，当前配置为服务商模式');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -39,7 +39,7 @@ class InvokeAndroidPluginTest extends TestCase
 
         self::expectException(InvalidResponseException::class);
         self::expectExceptionCode(Exception::RESPONSE_MISSING_NECESSARY_PARAMS);
-        self::expectExceptionMessage('Android调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求');
+        self::expectExceptionMessage('JSAPI调起用户确认收款失败：响应缺少 `package_info` 参数，请自行检查参数是否符合微信要求');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -47,26 +47,28 @@ class InvokeAndroidPluginTest extends TestCase
     public function testNormalParams()
     {
         $rocket = (new Rocket())
-            ->setDestination(new Collection(['package_info' => 'affffddafdfafddffda==']))
+            ->setDestination(new Collection(['package_info' => 'yansongda']))
             ->setPayload(['_invoke_appId' => '111']);
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         $contents = $result->getDestination();
 
-        self::assertEquals('requestMerchantTransfer', $contents->get('businessType'));
-        self::assertEquals('appId=111&mchId=1600314069&package=affffddafdfafddffda%3D%3D', $contents->get('query'));
+        self::assertEquals('111', $contents->get('appId'));
+        self::assertEquals('yansongda', $contents->get('package'));
+        self::assertEquals('1600314069', $contents->get('mchId'));
     }
 
     public function testNormal()
     {
-        $rocket = (new Rocket())->setDestination(new Collection(['package_info' => 'affffddafdfafddffda==']));
+        $rocket = (new Rocket())->setDestination(new Collection(['package_info' => 'yansongda']));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         $contents = $result->getDestination();
 
-        self::assertEquals('requestMerchantTransfer', $contents->get('businessType'));
-        self::assertEquals('appId=wx55955316af4ef13&mchId=1600314069&package=affffddafdfafddffda%3D%3D', $contents->get('query'));
+        self::assertEquals('wx55955316af4ef13', $contents->get('appId'));
+        self::assertEquals('yansongda', $contents->get('package'));
+        self::assertEquals('1600314069', $contents->get('mchId'));
     }
 }